Add block processing and rewrite audio interface
For some reason, audio stops working unless it is run within an async task or with `run_until_complete`. The program just stops without an error when processing network tasks. Notably, this doesn't occur when ran in a debugger, so I still have no idea why it might happen. Moving it into a callback seemed to have fixed it for now, but this is still an open problem.